+++ aiosmtpd/smtp.py
@@ -168,6 +168,16 @@ class SMTP(asyncio.StreamReaderProtocol):
         yield from self.push('250 %s' % self.hostname)
 
     @asyncio.coroutine
+    def ehlo_hook(self):
+        """Allow subclasses to extend EHLO responses.
+
+        This hook is called just before the final, non-continuing
+        `250 HELP` response.  Subclasses can add additional `250-<cmd>`
+        responses for custom behavior.
+        """
+        pass
+
+    @asyncio.coroutine
     def smtp_EHLO(self, arg):
         if not arg:
             yield from self.push('501 Syntax: EHLO hostname')
@@ -188,6 +198,7 @@ class SMTP(asyncio.StreamReaderProtocol):
         if self.enable_SMTPUTF8:
             yield from self.push('250-SMTPUTF8')
             self.command_size_limits['MAIL'] += 10
+        yield from self.ehlo_hook()
         yield from self.push('250 HELP')
 
     @asyncio.coroutine
@@ -392,11 +403,17 @@ class SMTP(asyncio.StreamReaderProtocol):
         yield from self.push('250 OK')
 
     @asyncio.coroutine
+    def rset_hook(self):
+        """Allow subclasses to hook into the RSET command."""
+        pass
+
+    @asyncio.coroutine
     def smtp_RSET(self, arg):
         if arg:
             yield from self.push('501 Syntax: RSET')
             return
         self._set_rset_state()
+        yield from self.rset_hook()
         yield from self.push('250 OK')
 
     @asyncio.coroutine